+++ lib/parser/models/config/android.g.dart
@@ -6,6 +6,17 @@ part of 'android.dart';
 // JsonSerializableGenerator
 // **************************************************************************
 
-Android _$AndroidFromJson(Map json) => Android(
-      flavorDimensions: json['flavorDimensions'] as String? ?? 'flavor-type',
-    );
+Android _$AndroidFromJson(Map json) {
+  $checkKeys(
+    json,
+    disallowNullValues: const ['resValues'],
+  );
+  return Android(
+    flavorDimensions: json['flavorDimensions'] as String? ?? 'flavor-type',
+    resValues: (json['resValues'] as Map?)?.map(
+          (k, e) => MapEntry(k as String,
+              ResValue.fromJson(Map<String, dynamic>.from(e as Map))),
+        ) ??
+        {},
+  );
+}