+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## NEXT
 
+* Added common resValues for Android
 * Changed the way flutter_flavorizr reads the instructions: from the flavorizr.yaml as a main source
   and pubspec.yaml as fallback
 * Android and iOS definitions are no longer required